# apps/stream_whip/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "**Whip** and **Whep** protocol implementation. `OBS =\u003e Server` - *Whip*. `Server \u003c=\u003e Client` (e.g from browser) - *Whep*. It's a **bridge** to make `WebRTC` connection.",
        "title": "WhipWhep API",
        "contact": {
            "name": "dehwyy",
            "url": "https://t.me/dehwyy",
            "email": "dehwyyy@gmail.icom"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/whep": {
            "post": {
                "description": "Receive **SdpOffer** and **AutorizationToken** from **Client**. Retrieve *Audio/Video* streams for future broadcoast via `WebRTC`.",
                "consumes": [
                    "application/sdp"
                ],
                "produces": [
                    "application/sdp"
                ],
                "tags": [
                    "WhipWhep"
                ],
                "summary": "*Client* and *Server* offers excahange.",
                "operationId": "HandleWhep",
                "parameters": [
                    {
                        "description": "*SDPOffer* generated by **Client**",
                        "name": "SDPOffer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Requested *audio+video* stream name",
                        "name": "X-StreamName",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ServerSdpAnswer",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/whep"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_dehwyy_mugen_apps_stream_whip_internal_server_errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_dehwyy_mugen_apps_stream_whip_internal_server_errors.HTTPError"
                        }
                    }
                }
            }
        },
        "/whip": {
            "post": {
                "description": "Receive **SdpOffer** and **AutorizationToken** from **OBS**. Create *Audio/Video* streams for future broadcoast via `WebRTC`.",
                "consumes": [
                    "application/sdp"
                ],
                "produces": [
                    "application/sdp"
                ],
                "tags": [
                    "WhipWhep"
                ],
                "summary": "*OBS* and *Server* offers exchange.",
                "operationId": "HandleWhip",
                "parameters": [
                    {
                        "description": "*SDPOffer* generated by **OBS**",
                        "name": "SDPOffer",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "type": "string",
                        "description": "AuthorizationToken set by **OBS**",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "ServerSdpAnswer",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "/whip"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_dehwyy_mugen_apps_stream_whip_internal_server_errors.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_dehwyy_mugen_apps_stream_whip_internal_server_errors.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_dehwyy_mugen_apps_stream_whip_internal_server_errors.HTTPError": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "err": {
                    "type": "string"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Essential router",
            "name": "WhipWhep"
        }
    ],
    "externalDocs": {
        "description": "IETF WHIP",
        "url": "https://datatracker.ietf.org/doc/html/draft-ietf-wish-whip-16"
    }
}